requirements.txt
# Core Dependencies
numpy<2.0.0
opencv-python>=4.5.0
setproctitle
python-dotenv

# GStreamer Python Bindings (for detection.py)
PyGObject>=3.36.0

# Kafka & Streaming (for kafka_handler.py)
kafka-python>=2.0.0

# AWS & Cloud Storage (for kafka_handler.py)
boto3>=1.20.0
botocore>=1.23.0

# Serial Communication (for radar_handler.py)
pyserial>=3.5

# Data Processing & Utilities
pyyaml>=5.4.0
requests>=2.25.0

# Testing & Development
pytest>=6.0.0
pytest-timeout
pytest-mock
pytest-asyncio

# Optional: Additional Computer Vision
pillow>=8.0.0

# Optional: Web Framework (if using FastAPI endpoints)
# fastapi>=0.68.0
# uvicorn[standard]>=0.15.0

# Optional: Video Processing (if using ffmpeg)
# ffmpeg-python
# imageio

# Optional: Data Analysis
# pandas>=1.3.0
# matplotlib

# Hailo-specific packages (installed separately by install.sh)
# hailort  # Installed by hailo_python_installation.sh
# tappas_core_python_binding  # Installed by hailo_python_installation.sh  
# hailo-apps  # Installed from Git by install.sh

# System packages required (install via apt/yum):
# python3-gi
# python3-gi-cairo
# gstreamer1.0-tools
# gstreamer1.0-plugins-base
# gstreamer1.0-plugins-good
# gstreamer1.0-plugins-bad
# gstreamer1.0-plugins-ugly